Fix enum and break up the cpp file into file per function
